# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "In Abstract Controller",
        "version": "1"
    },
    "paths": {
...
        "/admins/departments/{id}/close-for-registerations": {
            "patch": {
                "tags": [
                    "adminsDepartments"
                ],
                "operationId": "8130e63bf369f621c0e4dc36c8ea471e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CloseDepartmentForRegisterationRequestData"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "parameters": [
                {
                    "$ref": "#/components/parameters/CloseDepartmentForRegisterationPathParameterData"
                }
            ]
        },
...
        "/admins/departments/{id}/open-for-registerations": {
            "patch": {
                "tags": [
                    "adminsDepartments"
                ],
                "operationId": "d44fba78bfacc0914c14849237ea78cc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OpenDepartmentForRegisterationData"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "parameters": [
                {
                    "$ref": "#/components/parameters/adminsDepartmentIdPathParameterData"
                }
            ]
        },
...
        "/admins/course-teachers/classrooms": {
            "post": {
                "tags": [
                    "adminsCourseTeachers"
                ],
                "operationId": "584d4f967e794038f337d734cd0ee691",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AssignClassroomToCourseTeacherRequestData"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
        "/admins/course-teachers/exams": {
            "post": {
                "tags": [
                    "adminsCourseTeachers"
                ],
                "operationId": "51030059d7a266447d286423ac6274b3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateCourseTeacherExamRequestData"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/admins/courses": {
            "post": {
                "tags": [
                    "adminsCourses"
                ],
                "operationId": "a3763a3879fda6ba6dd4d7a1c56ba02d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateCourseRequestData"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "adminsCourses"
                ],
                "operationId": "dd364bf09bce2d6c9f8250b2d68f36a1",
                "parameters": [
                    {
                        "name": "ids[]",
                        "in": "query",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
        "/admins/open-course-registerations/teachers": {
            "post": {
                "tags": [
                    "adminsOpenCourseRegisterations"
                ],
                "operationId": "991d871503e411a02d9f5242097bfdd6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AssignTeacherToCourseRequestData"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
        "/admins/students": {
            "post": {
                "tags": [
                    "adminsStudents"
                ],
                "operationId": "34f3b35b8d3346312cd49329dc7d7a05",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterStudentRequestData"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
        "/students/courses/marks": {
            "get": {
                "tags": [
                    "studentsCourses"
                ],
                "operationId": "26e46e59cb305aa678b442cdb3e05c0f",
                "parameters": [
                    {
                        "name": "year",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "semester",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "perPage",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Items fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/GetCoursesMarksResponseData"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "AssignClassroomToCourseTeacherRequestData": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "classroom_id": {
                        "type": "integer"
                    },
                    "day": {
                        "type": "integer"
                    },
                    "from": {
                        "type": "string"
                    },
                    "to": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "CreateCourseTeacherExamRequestData": {
                "properties": {
                    "course_teacher_id": {
                        "type": "integer"
                    },
                    "classroom_id": {
                        "type": "integer"
                    },
                    "date": {
                        "schema": "string",
                        "type": "string",
                        "format": "datetime",
                        "default": "2024-09-02 18:31:45",
                        "pattern": "YYYY-MM-DD",
                        "example": "2024-09-02 18:31:45"
                    },
                    "from": {
                        "type": "string"
                    },
                    "to": {
                        "type": "string"
                    },
                    "max_mark": {
                        "type": "integer"
                    },
                    "is_main_exam": {
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
...
            "CreateCourseRequestData": {
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "code": {
                        "type": "string"
                    },
                    "is_active": {
                        "type": "boolean"
                    },
                    "credits": {
                        "type": "integer"
                    },
                    "department_id": {
                        "type": "integer",
                        "nullable": true
                    },
                    "prerequisites_ids": {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        }
                    }
                },
                "type": "object"
            },
...
            "CloseDepartmentForRegisterationRequestData": {
                "properties": {
                    "year": {
                        "type": "integer"
                    },
                    "semester": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
            "OpenDepartmentForRegisterationData": {
                "properties": {
                    "course_registeration_year": {
                        "type": "integer"
                    },
                    "course_registeration_semester": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
            "AssignTeacherToCourseRequestData": {
                "properties": {
                    "course_id": {
                        "type": "integer"
                    },
                    "teacher_id": {
                        "type": "integer"
                    },
                    "is_main_teacher": {
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
...
            "RegisterStudentRequestData": {
                "properties": {
                    "department_id": {
                        "type": "integer",
                        "nullable": true
                    },
                    "national_id": {
                        "type": "string"
                    },
                    "birthdate": {
                        "schema": "string",
                        "type": "string",
                        "format": "datetime",
                        "default": "2024-09-02 18:31:45",
                        "pattern": "YYYY-MM-DD",
                        "example": "2024-09-02 18:31:45"
                    },
                    "enrollment_date": {
                        "schema": "string",
                        "type": "string",
                        "format": "datetime",
                        "default": "2024-09-02 18:31:45",
                        "pattern": "YYYY-MM-DD",
                        "example": "2024-09-02 18:31:45"
                    },
                    "phone_number": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "CourseItemData": {
                "$ref": "GetCoursesMarksCourseItemData"
            },
            "GetCoursesMarksResponseData": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "course": {
                        "$ref": "#/components/schemas/CourseItemData"
                    },
                    "year": {
                        "type": "integer"
                    },
                    "semester": {
                        "type": "integer"
                    },
                    "final_mark": {
                        "type": "integer"
                    },
                    "exam_mark": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
        },
        "parameters": {
...
            "CloseDepartmentForRegisterationPathParameterData": {
                "name": "id",
                "in": "path",
                "required": true,
                "schema": {
                    "type": "integer"
                }
            },
            "adminsDepartmentIdPathParameterData": {
                "name": "id",
                "in": "path",
                "required": true,
                "schema": {
                    "type": "integer"
                }
            },
...
        },
        "securitySchemes": {
            "ref_to_be_used_below_csrf": {
                "type": "apiKey",
                "name": "X-XSRF-TOKEN",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "ref_to_be_used_below_csrf": []
        }
    ],
...
}